+++ tests/integration/transactions/test_did_set.py
@@ -4,7 +4,7 @@ from tests.integration.it_utils import (
     test_async_and_sync,
 )
 from tests.integration.reusable_values import WALLET
-from xrpl.models import AccountObjects, AccountObjectType, DIDSet
+from xrpl.models import AccountObjects, DIDSet, LedgerEntryType
 from xrpl.models.response import ResponseStatus
 
 _VALID_FIELD = "1234567890abcdefABCDEF"
@@ -25,6 +25,6 @@ class TestDIDSet(IntegrationTestCase):
 
         # confirm that the DID was actually created
         account_objects_response = await client.request(
-            AccountObjects(account=WALLET.address, type=AccountObjectType.DID)
+            AccountObjects(account=WALLET.address, type=LedgerEntryType.DID)
         )
         self.assertEqual(len(account_objects_response.result["account_objects"]), 1)