+++ tests/integration/transactions/test_set_oracle.py
@@ -6,7 +6,7 @@ from tests.integration.it_utils import (
     test_async_and_sync,
 )
 from tests.integration.reusable_values import WALLET
-from xrpl.models import AccountObjects, AccountObjectType, OracleSet
+from xrpl.models import AccountObjects, LedgerEntryType, OracleSet
 from xrpl.models.response import ResponseStatus
 from xrpl.models.transactions.oracle_set import PriceData
 from xrpl.utils import str_to_hex
@@ -54,7 +54,7 @@ class TestSetOracle(IntegrationTestCase):
 
         # confirm that the PriceOracle was actually created
         account_objects_response = await client.request(
-            AccountObjects(account=WALLET.address, type=AccountObjectType.ORACLE)
+            AccountObjects(account=WALLET.address, type=LedgerEntryType.ORACLE)
         )
 
         # subsequent integration tests (sync/async + json/websocket) add one